Move print_allocator_records to Umpire.{c,h}pp
