+++ src/umpire/Umpire.hpp
@@ -17,6 +17,10 @@
 
 #include "umpire/ResourceManager.hpp"
 
+#include "umpire/Allocator.hpp"
+
+#include <iostream>
+
 namespace umpire {
 
 /*!
@@ -46,6 +50,14 @@ void free(void* ptr)
 {
   return ResourceManager::getInstance().deallocate(ptr);
 }
+
+/*!
+ * \brief Print the allocations from a specific allocator
+ *
+ * \param allocator source Allocator.
+ * \param os output stream
+ */
+void print_allocator_records(Allocator allocator, std::ostream& os = std::cout);
 
 } // end of namespace umpire
